Fix various clang compiler errors for building with ft2232 support

+++ src/jtag/drivers/ft2232.c
@@ -1511,7 +1511,7 @@ static void minimodule_reset(int trst, int srst)
 
 static void turtle_reset(int trst, int srst)
 {
-	trst = trst;
+	trst = (int) trst;
 
 	if (srst == 1)
 		low_output |= nSRST;
